fix: made some enum to int casts explicit fix: use const ref arguments where it made more sense

+++ UDT_DLL/src/api.cpp
@@ -644,7 +644,7 @@ static bool CreateDemoFileSplit(udtVMLinearAllocator& tempAllocator, udtContext&
 	}
 
 	udtString newFilePath = udtString::NewEmpty(tempAllocator, UDT_MAX_PATH_LENGTH);
-	sprintf(newFilePath.GetWritePtr(), "%s_SPLIT_%u%s", outputFilePathStart.GetPtr(), index + 1, udtGetFileExtensionByProtocol(protocol));
+	sprintf(newFilePath.GetWritePtr(), "%s_SPLIT_%u%s", outputFilePathStart.GetPtr(), index + 1, udtGetFileExtensionByProtocol((u32)protocol));
 
 	context.LogInfo("Writing demo %s...", newFilePath.GetPtr());
